# openapi.yaml
openapi: 3.0.3
info:
  version: 0.1.0
  title: tapes
  description: |
    The **tapes** API porvides information about the tapes and images that have been
    synced to the tapes database.
externalDocs:
  description: 'github.com/golden-vcr/tapes'
  url: https://github.com/golden-vcr/tapes
servers:
  - url: https://goldenvcr.com/api/tapes
tags:
  - name: catalog
    description: |-
      Public endpoints that provide information about the tapes available for screening
      in the Golden VCR Library.
  - name: favorites
    description: |-
      Endpoints that allow an authenticated user to manage which tapes they've selected
      as their favorites.
paths:
  /catalog:
    get:
      tags:
        - catalog
      summary: |-
        Returns a listing of tapes
      operationId: getCatalog
      responses:
        '200':
          description: |-
            Catalog data was successfully fetched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogListing'
  /catalog/{tapeId}:
    get:
      tags:
        - catalog
      summary: |-
        Returns the details of a single tape
      parameters:
        - in: path
          name: tapeId
          schema:
            type: integer
          required: true
          description: Unique identifier for the tape to look up
          example: 13
      operationId: getCatalogItem
      responses:
        '200':
          description: |-
            Tape was found; details follow
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogItem'
        '404':
          description: |-
            No tape exists with the given ID
  /favorites:
    get:
      tags:
        - favorites
      summary: |-
        Returns the set of favorite tape IDs for the authenticated user
      security:
        - twitchUserAccessToken: []
      operationId: getFavorites
      responses:
        '200':
          description: |-
            Authentication OK; returning a set of 0 or more favorite tape IDs.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FavoriteTapeSet'
        '401':
          description: |-
            Authentication failed; caller's identity could not be ascertained.
    patch:
      tags:
        - favorites
      summary: |-
        Allows a single tape to be registered (or unregistered) as a favorite for the
        authenticated user
      security:
        - twitchUserAccessToken: []
      operationId: patchFavorites
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FavoriteTapeChange'
      responses:
        '204':
          description: |-
            OK; database state for the given tape and user now reflects the request.
        '400':
          description: |-
            Request refers to an invalid tape ID.
        '401':
          description: |-
            Authentication failed; caller's identity could not be ascertained.
components:
  schemas:
    CatalogListing:
      type: object
      properties:
        imageHost:
          type: string
          description: Base URL from which image URLs can be constructed
          example: https://golden-vcr-images.nyc3.digitaloceanspaces.com
        items:
          type: array
          description: Array of all tapes in the Golden VCR library
          items:
            $ref: '#/components/schemas/CatalogItem'
    CatalogItem:
      type: object
      properties:
        id:
          type: integer
          description: Unique identifier for the page
          example: 13
        title:
          type: string
          description: Title of the tape
          example: Build Your Own Computer Desk (American Plywood Association)
        year:
          type: integer
          description: Year of publication, or 0 if unknown
          example: 1987
        runtime:
          type: integer
          description: Approximate runtime in minutes, or 0 if unknown
          example: 25
        thumbnail:
          type: string
          description: Filename of thumbanil image, served relative to imageHost URL
          example: 0013_thumb.jpg
        images:
          type: array
          description: Array of one or more full-sized gallery images scanned from this tape
          items:
            $ref: '#/components/schemas/GalleryImage'
    GalleryImage:
      type: object
      properties:
        filename:
          type: string
          description: Filename of the image, served relative to imageHost URL
          example: 0013_a.jpg
        width:
          type: integer
          description: Width of the image in pixels
          example: 624
        height:
          type: integer
          description: Height of the image in pixels
          example: 1116
        color:
          type: string
          description: Hex string representing the dominant color in the image
          example: '#c9ab86'
        rotated:
          type: boolean
          description: |
            Whether the image was rotated 90 degrees CCW in order to have a vertical
            aspect ratio, in which case it may be rotated 90 degrees CW to be displayed
            with the text upright
          example: false
    FavoriteTapeSet:
      type: object
      properties:
        tapeIds:
          type: array
          description: List of every tape ID which the user has marked as a favorite
          items:
            type: integer
          example: [1, 3, 42]
    FavoriteTapeChange:
      type: object
      properties:
        tapeId:
          type: integer
          example: 44
        isFavorite:
          type: boolean
  securitySchemes:
    twitchUserAccessToken:
      type: http
      scheme: bearer